+++ apps/mis-server/src/entities/UserAccount.ts
@@ -10,16 +10,25 @@
  * See the Mulan PSL v2 for more details.
  */
 
-import { Entity, ManyToOne, PrimaryKey, Property, Ref } from "@mikro-orm/core";
+import { Entity, Enum, ManyToOne, PrimaryKey, Property, Ref } from "@mikro-orm/core";
 import { Decimal } from "@scow/lib-decimal";
 import { Account } from "src/entities/Account";
 import { User } from "src/entities/User";
 import { DecimalType } from "src/utils/decimal";
 import { EntityOrRef, toRef } from "src/utils/orm";
 
+// 用户在集群中的状态
 export enum UserStatus {
   UNBLOCKED = "UNBLOCKED",
   BLOCKED = "BLOCKED",
+}
+
+// 用户在账户中的状态，是否被手动封锁
+export enum UserStateInAccount {
+  // 未被手动封锁
+  NORMAL = "NORMAL",
+  // 被上级手动封锁
+  BLOCKED_BY_ADMIN = "BLOCKED_BY_ADMIN",
 }
 
 export enum UserRole {
@@ -40,7 +49,7 @@ export class UserAccount {
     account: Ref<Account>;
 
   @Property({ columnType: "varchar(10)", comment: Object.values(UserStatus).join(", ") })
-    status: UserStatus;
+    blockedInCluster: UserStatus;
 
   @Property({ columnType: "varchar(10)", comment: Object.values(UserRole).join(", ") })
     role: UserRole;
@@ -51,19 +60,25 @@ export class UserAccount {
   @Property({ type: DecimalType, nullable: true })
     jobChargeLimit?: Decimal;
 
+  @Enum({ items: () => UserStateInAccount,
+    default: UserStateInAccount.NORMAL, comment: Object.values(UserStateInAccount).join(", ") })
+    state?: UserStateInAccount;
+
   constructor(init: {
     user: EntityOrRef<User>,
     account: EntityOrRef<Account>,
-    status: UserStatus,
+    blockedInCluster: UserStatus,
     role: UserRole,
     jobChargeLimit?: Decimal,
     usedJobCharge?: Decimal,
+    state?: UserStateInAccount,
   }) {
     this.user = toRef(init.user);
     this.account = toRef(init.account);
-    this.status = init.status;
+    this.blockedInCluster = init.blockedInCluster;
     this.role = init.role;
     this.jobChargeLimit = init.jobChargeLimit;
     this.usedJobCharge = init.usedJobCharge;
+    this.state = init.state || UserStateInAccount.NORMAL;
   }
 }