+++ apps/mis-server/tests/data/data.ts
@@ -43,19 +43,22 @@ export async function insertInitialData(em: SqlEntityManager) {
   const uaAA = new UserAccount({
     account: accountA,
     user: userA,
-    role: UserRole.OWNER, status: UserStatus.UNBLOCKED,
+    role: UserRole.OWNER,
+    blockedInCluster: UserStatus.UNBLOCKED,
   }) as Loaded<UserAccount, "account" | "user">;
 
   const uaAB = new UserAccount({
     account: accountA,
     user: userB,
-    role: UserRole.ADMIN, status: UserStatus.UNBLOCKED,
+    role: UserRole.ADMIN,
+    blockedInCluster: UserStatus.UNBLOCKED,
   }) as Loaded<UserAccount, "account" | "user">;
 
   const uaBB = new UserAccount({
     account: accountB,
     user: userB,
-    role: UserRole.OWNER, status: UserStatus.UNBLOCKED,
+    role: UserRole.OWNER,
+    blockedInCluster: UserStatus.UNBLOCKED,
   }) as Loaded<UserAccount, "account" | "user">;
 
   await em.persistAndFlush([userA, userB, accountA, accountB, uaAA, uaAB, uaBB]);
@@ -69,8 +72,12 @@ export async function insertInitialData(em: SqlEntityManager) {
     blockedInCluster: false,
     comment: "123",
   });
-  const uaCC = new UserAccount({ user: userC, account: accountC, role: UserRole.ADMIN, status: UserStatus.BLOCKED });
-
+  const uaCC = new UserAccount({
+    user: userC,
+    account: accountC,
+    role: UserRole.ADMIN,
+    blockedInCluster: UserStatus.BLOCKED,
+  });
   await em.persistAndFlush([anotherTenant, userC, accountC, uaCC]);
 
   return { tenant, userA, userB, userC, accountA, accountB, accountC, uaAA, uaAB, uaBB, uaCC, anotherTenant };
@@ -104,19 +111,22 @@ export async function insertBlockedData(em: SqlEntityManager) {
   const uaAA = new UserAccount({
     account: unblockedAccountA,
     user: blockedUserA,
-    role: UserRole.OWNER, status: UserStatus.BLOCKED,
+    role: UserRole.OWNER,
+    blockedInCluster: UserStatus.BLOCKED,
   });
 
   const uaAB = new UserAccount({
     account: unblockedAccountA,
     user: unblockedUserB,
-    role: UserRole.ADMIN, status: UserStatus.UNBLOCKED,
+    role: UserRole.ADMIN,
+    blockedInCluster: UserStatus.UNBLOCKED,
   });
 
   const uaBB = new UserAccount({
     account: blockedAccountB,
     user: unblockedUserB,
-    role: UserRole.OWNER, status: UserStatus.UNBLOCKED,
+    role: UserRole.OWNER,
+    blockedInCluster: UserStatus.UNBLOCKED,
   });
 
   await em.persistAndFlush([uaAA, uaAB, uaBB]);